Added basic storyboard, changed the GoogleService-Info.json file and added a .gitignore
Realised that I had accidentally added the wrong GoogleService-Info.json file to the project (the curse of never closing tabs). Fixed that, but also added a .gitignore to prevent the file from being added to GitHub - the file contains sensitive keys, so shouldn't be publicly available. I also created a basic storyboard layout, with most of the Welcome Page completed